test(Timeline): Disable smooth scrolling in Cypress
The Timeline tests have still been flakey on GitHub Actions.

This disables smooth scrolling for them in Cypress to hopefully fix it. (Smooth scrolling didn't make much sense here either during the tests as the tests check if something was scrolled to, so you could get a false positive if it scrolls past the target.)